requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
demjson3==3.0.6
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
docstring-parser==0.15
exceptiongroup==1.2.0
FLAML==2.1.1
frozenlist==1.4.0
fsspec==2023.12.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
joblib==1.3.2
llama-index==0.9.10
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
openai==1.2.4
packaging==23.2
pandas==2.1.4
prettytable==3.9.0
prompt-toolkit==3.0.36
pyautogen==0.2.1
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
Pygments==2.17.2
pymemgpt==0.2.6
pypdf==3.17.1
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
